Extract section_heading + add_border_to_ascii_art into modules/output_headers.py
Ninth refactor pass at modules/output.py.  Two small ASCII-art helpers
moved into a dedicated module.  Kometa config files carry decorative
section headers between YAML sections -- either single-line comments
or pyfiglet-generated ASCII art wrapped in a hash-mark border.

Functions moved:
  * add_border_to_ascii_art  -- wrap a pyfiglet block with '#'-borders
  * section_heading          -- render a title in the chosen font style,
    fallback to '===...=== title ===...===' on 'none'/'single line'/
    unknown-font

Minor cleanup: section_heading's elif/else chain rewritten with early
returns (Zen of Python: 'Flat is better than nested').

Public surface: quickstart.py calls output.section_heading directly
(line 1299) so this MUST re-export it.  output.py imports both names
without underscore, no # noqa needed.

Note: pyfiglet is still imported in output.py because six other
callers use pyfiglet.figlet_format directly inline (mostly duplicated
section_heading-like logic in build_config / build_libraries_section).
Future DRY target: replace those inline copies with section_heading()
calls once the mega-functions are broken up.

Impact:
- modules/output.py: 2839 -> 2822 (-17 / -0.6%)
- modules/output_headers.py: 37 lines new

Cumulative on output.py: 3833 -> 2822 (-1011 / -26.4%)  -- BROKE -1000!

All 1077 tests pass. Ruff + black clean.
